requirements.txt
tensorflow
fastapi
uvicorn[standard]
python-multipart
numpy
Pillow
pillow-heif